+++ controllers/finish_type.controller.js
@@ -12,9 +12,9 @@ exports.getAll = async (req, res) => {
 };
 
 exports.create = async (req, res) => {
-  const { name } = req.body;
+  const { name, SurfaceId } = req.body;
   try {
-    await FinishType.create({ name, image: req.file.filename });
+    await FinishType.create({ name, image: req.file.filename,SurfaceId });
     return res
       .status(200)
       .json({ success: true, message: "Project Type created successfully" });
@@ -41,6 +41,9 @@ exports.getSpecificFinish = async (req, res) => {
 exports.update = async (req, res) => {
   const updateFinishType = req.body;
   const { id } = req.params;
+  if (!id) {
+    throw "Id is missing or incorrect format";
+  }
   if (req.file) {
     updateFinishType['image'] = req.file.filename;
     const { image } = await FinishType.find({ where: { id: req.params.id }, raw: true });
@@ -64,6 +67,9 @@ exports.delete = async (req, res) => {
   const { image } = await FinishType.find({ where: { id: id }, raw: true });
   fs.unlinkSync(`${keys.storage}/${image}`);
   try {
+    if (!id) {
+      throw "Id is missing or incorrect format";
+    }
     await FinishType.destroy({
       where: {
         id
